+++ source/plugins/projects.editor.opcuamodel/visual/language_definitions.js
@@ -8,9 +8,8 @@ module.exports = function (blockly) {
 	Blockly.Blocks['opcuamodel_variable'] = {
 		init: function () {
 			this.appendDummyInput()
-				.appendField('OPC/UA Variable');
+				.appendField('Variable');
 			this.appendDummyInput()
-				.appendField('Name')
 				.appendField(new Blockly.FieldTextInput('variable_name'), 'variable_name');
 			this.appendDummyInput()
 				.appendField('Data Type')
@@ -40,7 +39,7 @@ module.exports = function (blockly) {
 			this.appendDummyInput()
 				.appendField('History')
 				.appendField(new Blockly.FieldCheckbox('FALSE'), 'variable_history');
-			this.setInputsInline(false);
+			this.setInputsInline(true);
 			this.setPreviousStatement(true, ['opcuamodel_property', 'opcuamodel_object', 'opcuamodel_folder']);
 			this.setNextStatement(true, ['opcuamodel_property', 'opcuamodel_object', 'opcuamodel_folder']);
 			this.setColour(230);
@@ -52,13 +51,28 @@ module.exports = function (blockly) {
 	Blockly.Blocks['opcuamodel_folder'] = {
 		init: function () {
 			this.appendDummyInput()
-				.appendField('OPC/UA Folder')
+				.appendField('Folder')
 				.appendField(new Blockly.FieldTextInput('folder_name'), 'folder_name');
 			this.appendStatementInput('folder_objects_properties')
 				.setCheck(['opcuamodel_property', 'opcuamodel_object', 'opcuamodel_folder']);
 			this.setPreviousStatement(true, ['opcuamodel_object', 'opcuamodel_folder']);
 			this.setNextStatement(true, ['opcuamodel_object', 'opcuamodel_folder']);
-			this.setColour(230);
+			this.setColour(210);
+			this.setTooltip('');
+			this.setHelpUrl('');
+		}
+	};
+
+	Blockly.Blocks['opcuamodel_template'] = {
+		init: function () {
+			this.appendDummyInput()
+				.appendField('Folder Template')
+				.appendField(new Blockly.FieldTextInput('template_name'), 'template_name');
+			this.appendStatementInput('folder_objects_properties')
+				.setCheck(['opcuamodel_property', 'opcuamodel_object', 'opcuamodel_folder']);
+			this.setPreviousStatement(true, ['opcuamodel_object', 'opcuamodel_folder']);
+			this.setNextStatement(true, ['opcuamodel_object', 'opcuamodel_folder']);
+			this.setColour(220);
 			this.setTooltip('');
 			this.setHelpUrl('');
 		}