+++ source/plugins/projects.editor.visual/visual/language_definitions.js
@@ -444,5 +444,5 @@ module.exports = function (blockly) {
 			this.setTooltip('');
 		}
 	};
-
+	
 };