+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## Wyliodrin STUDIO 2.1.1-beta
+  * **added** folder template for opcua model
+  * **fixed** multiple instance popup appearing multiple times
+  * **fixed** gutter background color in electron
+  * **fixed** image error from RPK simulator
+  * **fixed** display of extra brace at device os version
+  * **fixed** monaco editor workers build in the plugin folder
+  * **update** electron to version 7
+  * **update** serialport to version 8
+  * **update** new color for opcua model folder
+
+## Wyliodrin STUDIO 2.0.10-beta
+
+  * **added** social (email, facebook twitter and twilio) visual programming blocks
+  * **added** Adafruit DHT sensor visual programming blocks for Raspberry Pi and BeagleBone Black
+  * **added** delay block to visual programming blocks
+  * **feature** Wyapp PackageManager search with any case
+  * **fixed** Wyapp FileManager displaying double names
+  * **fixed** (partially) scroll bug in Firefox (temporary fix)
+
 ## Wyliodrin STUDIO 2.0.9-beta
 
 *First Release*